+++ test/test_util.py
@@ -7,6 +7,20 @@ def test_flatten_dictionary():
     assert x == {'a': 1, 'b.c': 2}
 
 
+def test_flatten_dictionary_with_ignores():
+    d = {'no_flatten': {'a': 1}, 'flatten': {'a': 1}}
+    x = flatten_dictionary(d, ignore_keys={'no_flatten'})
+    assert x['no_flatten']['a'] == 1
+    assert 'flatten.a' in x
+
+
+def test_flatten_dictionary_with_ignores_nested():
+    d = {'no': {'no_flatten': {'a': 1}}, 'flatten': {'a': 1}}
+    x = flatten_dictionary(d, ignore_keys={'no.no_flatten'})
+    assert x['no.no_flatten']['a'] == 1
+    assert 'flatten.a' in x
+
+
 def test_add_suffix():
     import pathlib
     p = pathlib.Path('foo.bar')